--- a/nucliadb/nucliadb/search/search/fetch.py
+++ b/nucliadb/nucliadb/search/search/fetch.py
@@ -25,7 +25,8 @@ from nucliadb_protos.resources_pb2 import Paragraph
 
 from nucliadb.ingest.orm.resource import KB_REVERSE
 from nucliadb.ingest.orm.resource import Resource as ResourceORM
-from nucliadb.ingest.serialize import serialize
+from nucliadb.ingest.serialize import managed_serialize
+from nucliadb.ingest.txn_utils import get_transaction
 from nucliadb.search import SERVICE_NAME, logger
 from nucliadb_models.common import FieldTypeName
 from nucliadb_models.resource import ExtractedDataTypeName, Resource
@@ -46,8 +47,10 @@ async def fetch_resources(
     extracted: list[ExtractedDataTypeName],
 ) -> dict[str, Resource]:
     result = {}
+    txn = await get_transaction(read_only=True)
     for resource in resources:
-        serialization = await serialize(
+        serialization = await managed_serialize(
+            txn,
             kbid,
             resource,
             show,