### Description
The way resources were serialized when crafting search results was creating too many pg txns and quickly exhausting the pol, therefore causing a performance bottleneck.

### How was this PR tested?
Local performance tests:
Before
```
Metrics summary:
- find_latency_s -> p50: 352.24ms p95: 515.18ms
- search_latency_s -> p50: 2.19s p95: 2.94s
- suggest_latency_s -> p50: 302.58ms p95: 900.45ms
==================================================
Errors summary:
==================================================
**** Molotov v2.6. Happy breaking! ****
SUCCESSES: 90 | FAILURES: 0
```

After:
```
Metrics summary:
- find_latency_s -> p50: 86.31ms p95: 159.79ms
- search_latency_s -> p50: 343.13ms p95: 493.50ms
- suggest_latency_s -> p50: 99.34ms p95: 218.76ms
==================================================
Errors summary:
==================================================
**** Molotov v2.6. Happy breaking! ****
SUCCESSES: 403 | FAILURES: 0
*** Bye ***
```

This means approximately a 4x improvement